mt5_bridge/requirements.txt
MetaTrader5==5.0.4424
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
loguru==0.7.3
